+++ src/main/kotlin/io/github/pandier/intellijdiscordrp/activity/ActivityFactory.kt
@@ -1,7 +1,8 @@
 package io.github.pandier.intellijdiscordrp.activity
 
-import de.jcm.discordgamesdk.activity.Activity
 import io.github.pandier.intellijdiscordrp.settings.ImageSetting
+import io.github.vyfor.kpresence.rpc.Activity
+import io.github.vyfor.kpresence.rpc.activity
 
 private fun ImageSetting.getIcon(context: ActivityContext) = when (this) {
     ImageSetting.APPLICATION -> currentActivityApplicationType.icon
@@ -18,23 +19,27 @@ class ActivityFactory(
     private val smallImageText: String = "",
     private val timestampEnabled: Boolean = true,
 ) {
-    fun create(context: ActivityContext): Activity = Activity().also {
-        if (details.isNotBlank())
-            it.details = displayMode.format(details, context)
-        if (state.isNotBlank())
-            it.state = displayMode.format(state, context)
+    fun create(context: ActivityContext): Activity = activity {
+        if (this@ActivityFactory.details.isNotBlank())
+            details = displayMode.format(this@ActivityFactory.details, context)
+        if (this@ActivityFactory.state.isNotBlank())
+            state = displayMode.format(this@ActivityFactory.state, context)
 
-        if (largeImage != null && largeImageText.isNotBlank()) {
-            it.assets().largeImage = largeImage.getIcon(context)
-            it.assets().largeText = displayMode.format(largeImageText, context)
+        assets {
+            if (this@ActivityFactory.largeImage != null && this@ActivityFactory.largeImageText.isNotBlank()) {
+                largeImage = this@ActivityFactory.largeImage.getIcon(context)
+                largeText = displayMode.format(this@ActivityFactory.largeImageText, context)
+            }
+
+            if (this@ActivityFactory.smallImage != null && this@ActivityFactory.smallImageText.isNotBlank()) {
+                smallImage = this@ActivityFactory.smallImage.getIcon(context)
+                smallText = displayMode.format(this@ActivityFactory.smallImageText, context)
+            }
         }
 
-        if (smallImage != null && smallImageText.isNotBlank()) {
-            it.assets().smallImage = smallImage.getIcon(context)
-            it.assets().smallText = displayMode.format(smallImageText, context)
+        timestamps {
+            if (timestampEnabled)
+                start = context.start.toEpochMilli()
         }
-
-        if (timestampEnabled)
-            it.timestamps().start = context.start
     }
 }