+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.0] - 2024-05-02
+
 ### Added
 
 - "Show elapsed time" checkbox that toggles the timestamp for a display mode
@@ -71,7 +73,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.1.0...v1.1.1
 [1.1.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.0.2...v1.1.0
 [1.0.2]: https://github.com/pandier/intellij-discord-rp/compare/v1.0.1...v1.0.2